requirements.txt
tweepy
kafka-python
boto3
pandas
delta-spark